test(bulk-import): stop replacing src/logger.ts process-wide with mock.module (issue 864)
scripts/__tests__/bulk-import.test.ts replaced src/logger.ts with a stub
object through `mock.module`. That mock is process-wide and permanent in
Bun -- keyed by resolved specifier, never scoped to one file, and not undone
by `mock.restore()` -- so once this file had loaded, every later suite in the
run received the stub instead of the real logger.

The visible failure: a suite that registers an `addLogSink` observer and
asserts on the lines the logger emits captures nothing, because the stub has
no sink machinery and pushes into this file's own arrays instead. The suite
passes alone and fails in the full run, with a stack pointing nowhere near
this file.

src/observability/observability.test.ts already states that no suite may
mock.module the logger, and scripts/backfill.test.ts is the precedent for the
alternative: observe through `addLogSink`, which composes with other suites
instead of overwriting them. This file now does the same, and the two
assertions it makes on `logCalls.warn` are unchanged.

Bringing the file to the whole-file lint standard came with it, since the
pre-commit gate lints staged files whole and this one carried 38 findings at
main (M14 inherited lint debt):

- 35 non-null assertions removed: `params![n]` loses the assertion the
  surrounding `as [string, unknown[]]` already makes redundant, and the three
  `files[0]!` / `warnCall![1]!` sites use `expectDefined` from
  scripts/test-support/expect-defined.ts.
- The file was 816 code lines against a 500 standard, so it splits by
  responsibility into scripts/__tests__/bulk-import.test.ts (parsing, reading,
  and the small/medium/large import paths) and
  scripts/__tests__/bulk-import-routing.test.ts (table routing, tag handling,
  dry run, argument parsing, edge cases), over a shared
  scripts/__tests__/bulk-import-fixture.ts holding the module mocks, the mock
  pool, and the log observer both need.
- Two describe callbacks ran past the 100-line function standard and split at
  a responsibility seam into "table routing" / "table routing: sessions" and
  "edge cases" / "edge cases: error logging and hashing". No `it` name, body,
  order, or assertion changed; neither file is a *.pg.test.ts, so
  scripts/assert-db-tests-ran.ts is untouched.

Behavior is unchanged: 59 tests before, 59 after, same names.

- Done-means: bun run test:isolated scripts/__tests__/bulk-import.test.ts src/tools/__tests__/agent-context-pack-durable-memory-recall-failure.test.ts

Refs issue 864.
